+++ packages/typegpu/src/typedDevice.ts
